auth işlemleri movie servise entegre edildi

+++ MovieApp.Service/Services/Concrete/MovieService.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using AutoMapper;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using MovieApp.Data.Context;
 using MovieApp.Data.UnitOfWorks;
 using MovieApp.Entity.Entities;
@@ -14,20 +15,32 @@ namespace MovieApp.Service.Services.Concrete
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
         private readonly IHttpContextAccessor _httpContextAccessor;
-        private readonly ClaimsPrincipal _user;
         private readonly ApplicationDbContext _dbContext;
+        private readonly UserManager<AppUser> _userManager;
 
-        public MovieService(IUnitOfWork unitOfWork, IMapper mapper, IHttpContextAccessor httpContextAccessor, ApplicationDbContext dbContext)
+        public MovieService(IUnitOfWork unitOfWork, IMapper mapper, IHttpContextAccessor httpContextAccessor, ApplicationDbContext dbContext, UserManager<AppUser> userManager)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _httpContextAccessor = httpContextAccessor;
-            _user = httpContextAccessor.HttpContext.User;
             _dbContext = dbContext;
+            _userManager = userManager;
         }
 
-        public async Task CreateAsync(MovieAddDto movieAddDto)
+        public async Task CreateAsync(MovieAddDto movieAddDto, int appUserId)
         {
+            var currentUser = await _userManager.FindByIdAsync(appUserId.ToString());
+
+            if (currentUser == null)
+            {
+                throw new Exception("Kullanýcý bulunamadý.");
+            }
+
+            if (!_userManager.IsInRoleAsync(currentUser, "Admin").Result)
+            {
+                throw new UnauthorizedAccessException("Bu iþlemi yapmak için yetkiniz yok.");
+            }
+
             Movie movie = new(
                 movieAddDto.Title,
                 movieAddDto.Description,
@@ -40,8 +53,20 @@ namespace MovieApp.Service.Services.Concrete
             await _unitOfWork.SaveAsync();
         }
 
-        public async Task<string> DeleteAsync(int movieId)
+        public async Task<string> DeleteAsync(int movieId , int appUserId)
         {
+            var currentUser = await _userManager.FindByIdAsync(appUserId.ToString());
+
+            if (currentUser == null)
+            {
+                throw new Exception("Kullanýcý bulunamadý.");
+            }
+
+            if (!_userManager.IsInRoleAsync(currentUser, "Admin").Result)
+            {
+                throw new UnauthorizedAccessException("Bu iþlemi yapmak için yetkiniz yok.");
+            }
+
             var movie = await _unitOfWork.GetRepository<Movie>().GetByGuidAsync(movieId);
 
             await _unitOfWork.GetRepository<Movie>().DeleteAsync(movie);
@@ -50,22 +75,49 @@ namespace MovieApp.Service.Services.Concrete
             return movie.Title;
         }
 
-        public async Task<List<MovieDto>> GetAllAsync()
+        public async Task<List<MovieDto>> GetAllAsync(int appUserId)
         {
+            var currentUser = await _userManager.FindByIdAsync(appUserId.ToString());
+
+            if (currentUser == null)
+            {
+                throw new Exception("Kullanýcý bulunamadý.");
+            }
+
             var movies = await _unitOfWork.GetRepository<Movie>().GetAllAsync();
             var map = _mapper.Map<List<MovieDto>>(movies);
 
             return map;
         }
 
-        public async Task<Movie> GetById(int id)
+        public async Task<Movie> GetById(int id, int appUserId)
         {
+            var currentUser = await _userManager.FindByIdAsync(appUserId.ToString());
+
+            if (currentUser == null)
+            {
+                throw new Exception("Kullanýcý bulunamadý.");
+            }
+
             var movie = await _unitOfWork.GetRepository<Movie>().GetByGuidAsync(id);
             return movie;
         }
 
-        public async Task<string> UpdateAsync(MovieUpdateDto movieUpdateDto)
+        public async Task<string> UpdateAsync(MovieUpdateDto movieUpdateDto, int appUserId)
         {
+
+            var currentUser = await _userManager.FindByIdAsync(appUserId.ToString());
+
+            if (currentUser == null)
+            {
+                throw new Exception("Kullanýcý bulunamadý.");
+            }
+
+            if (!_userManager.IsInRoleAsync(currentUser, "Admin").Result)
+            {
+                throw new UnauthorizedAccessException("Bu iþlemi yapmak için yetkiniz yok.");
+            }
+
             var movie = await _unitOfWork.GetRepository<Movie>().GetAsync(x => x.Id == movieUpdateDto.Id);
 
             movie.Title = movieUpdateDto.Title;
